# Conflicts:
#	stp-ms-lessons/src/main/java/com/sytoss/lessons/convertors/PumlConvertor.java
#	stp-ms-lessons/src/test/java/com/sytoss/lessons/bdd/common/TestExecutionContext.java

+++ stp-domain/src/main/java/com/sytoss/domain/bom/personalexam/Grade.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class Grade {
 
     @JsonView(PersonalExam.Public.class)
-    private float value;
+    private double value;
 
     @JsonView(PersonalExam.Public.class)
     private String comment;
